Move Mondrian classes to lib/rubiks/mondrian
