+++ lib/rubiks/mondrian/errors.rb
